UpgradeHandler for "Radon" Upgrade (#154)
