+++ .github/workflows/docker-publish-and-E2E.yaml
@@ -119,7 +119,7 @@ jobs:
       - name: checkout chain
         uses: actions/checkout@v2
         
-      - run: make ictest-paramauthorityChainUpgrade
+      - run: make ictest-chain-upgrade
         env:
           BRANCH_CI: ${{needs.build-and-push-image.outputs.branchTag}}
