+++ app/app.go
@@ -88,7 +88,8 @@ import (
 	"github.com/cosmos/cosmos-sdk/x/staking"
 	stakingkeeper "github.com/cosmos/cosmos-sdk/x/staking/keeper"
 	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
-	"github.com/strangelove-ventures/noble/app/upgrades/neon"
+	neon "github.com/strangelove-ventures/noble/app/upgrades/neon"
+	radon "github.com/strangelove-ventures/noble/app/upgrades/radon"
 	"github.com/strangelove-ventures/noble/cmd"
 	"github.com/strangelove-ventures/noble/docs"
 	"github.com/strangelove-ventures/noble/x/blockibc"
@@ -96,6 +97,7 @@ import (
 	fiattokenfactorymodulekeeper "github.com/strangelove-ventures/noble/x/fiattokenfactory/keeper"
 	fiattokenfactorymoduletypes "github.com/strangelove-ventures/noble/x/fiattokenfactory/types"
 	"github.com/strangelove-ventures/noble/x/globalfee"
+	globalfeetypes "github.com/strangelove-ventures/noble/x/globalfee/types"
 	tariff "github.com/strangelove-ventures/noble/x/tariff"
 	tariffkeeper "github.com/strangelove-ventures/noble/x/tariff/keeper"
 	tarifftypes "github.com/strangelove-ventures/noble/x/tariff/types"
@@ -843,6 +845,16 @@ func (app *App) setupUpgradeHandlers() {
 			app.BankKeeper,
 			app.AccountKeeper))
 
+	// radon upgrade
+	app.UpgradeKeeper.SetUpgradeHandler(
+		radon.UpgradeName,
+		radon.CreateRadonUpgradeHandler(
+			app.mm,
+			app.configurator,
+			app.ParamsKeeper,
+			app.FiatTokenFactoryKeeper,
+		))
+
 	upgradeInfo, err := app.UpgradeKeeper.ReadUpgradeInfoFromDisk()
 	if err != nil {
 		panic(fmt.Errorf("failed to read upgrade info from disk: %w", err))
@@ -857,6 +869,10 @@ func (app *App) setupUpgradeHandlers() {
 	case neon.UpgradeName:
 		stroreUpgrades = &storetypes.StoreUpgrades{
 			Added: []string{fiattokenfactorymoduletypes.StoreKey},
+		}
+	case radon.UpgradeName:
+		stroreUpgrades = &storetypes.StoreUpgrades{
+			Added: []string{globalfeetypes.ModuleName, tarifftypes.ModuleName},
 		}
 	}
