+++ app/upgrades/radon/constants.go
@@ -1,0 +1,8 @@
+package radon
+
+const (
+	// UpgradeName is the shared upgrade plan name for mainnet
+	UpgradeName = "radon"
+	// UpgradeInfo defines the binaries that will be used for the upgrade
+	UpgradeInfo = "Noble Radon Upgrade"
+)
